Implemented Input Scaling
Implemented a scaler which can be used to scale the bounds before passing to the sampler. If the scaler is also passed to the sampler, x_new values are scaled before being passed to the generator. Also implemented neural foil moment coefficient database and got reasonable results for central peak comparisons between neural nets and gaussian processes.